requirements.txt
pandas>=2.2
openpyxl>=3.1
numpy>=1.26
faiss-cpu>=1.8
requests>=2.32
python-dotenv>=1.0
